refactor: move FormProjection from core to quillmark crate
FormProjection is a consumer-facing convenience for form editors — it
composes existing QuillConfig APIs but is never used inside core. Move
it to the engine-layer `quillmark` crate so core stays focused on the
parse/schema/backend primitives.

- Relocate types and logic to `crates/quillmark/src/form.rs`
- Replace `impl Quill { project_form }` with a free function
  `quillmark::form::project_form(&Quill, &Document) -> FormProjection`
  (inherent impls can't cross crates)
- Move the project_form tests to the quillmark crate
- Update python and wasm bindings to call `quillmark::form::project_form`
