+++ crates/core/src/quill.rs
@@ -1,7 +1,6 @@
 //! Quill format bundle types and implementations.
 
 mod config;
-pub mod form;
 mod formats;
 mod ignore;
 mod load;
@@ -13,7 +12,6 @@ mod types;
 pub(crate) mod validation;
 
 pub use config::{CoercionError, QuillConfig};
-pub use form::{FormCard, FormFieldSource, FormFieldValue, FormProjection};
 pub use ignore::QuillIgnore;
 pub use tree::FileTreeNode;
 pub use types::{
