+++ crates/quillmark/src/lib.rs
@@ -23,8 +23,12 @@ pub use quillmark_core::{
     Severity,
 };
 
-// Declare orchestration module
+// Declare modules
+pub mod form;
 pub mod orchestration;
+
+// Re-export commonly-used form types at the crate root
+pub use form::{FormCard, FormFieldSource, FormFieldValue, FormProjection};
 
 // Re-export types from orchestration module
 pub use orchestration::{QuillRef, Quillmark, Workflow};
